+++ backend/src/main/java/ru/mephi/atomSkills25/commandA/choiceGame/repositories/impl/GameRepositoryImpl.java
@@ -58,7 +58,7 @@ public class GameRepositoryImpl implements GameRepository {
         }
     }
 
-    public void deleteGame(Long id) {
+    public void deleteGame(Integer id) {
         try {
             jdbcTemplate.update(DELETE_GAME, id);
         } catch (DataAccessException e) {
