+++ backend/src/main/java/ru/mephi/atomSkills25/commandA/choiceGame/services/PlayerService.java
@@ -10,5 +10,5 @@ public interface PlayerService {
 
     Integer addPlayer(NewPlayer newPlayer);
 
-    void deletePlayer(Long id);
+    void deletePlayer(Integer id);
 }
